Move appendVolumeIfNotExists method to utils package
