+++ pkg/controller/inferenceservice/resources/credentials/service_account_credentials.go
@@ -24,6 +24,7 @@ import (
 	"github.com/kubeflow/kfserving/pkg/controller/inferenceservice/resources/credentials/azure"
 	"github.com/kubeflow/kfserving/pkg/controller/inferenceservice/resources/credentials/gcs"
 	"github.com/kubeflow/kfserving/pkg/controller/inferenceservice/resources/credentials/s3"
+	"github.com/kubeflow/kfserving/pkg/utils"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -101,7 +102,7 @@ func (c *CredentialBuilder) CreateSecretVolumeAndEnv(namespace string, serviceAc
 		} else if _, ok := secret.Data[gcsCredentialFileName]; ok {
 			log.Info("Setting secret volume for gcs", "GCSSecret", secret.Name)
 			volume, volumeMount := gcs.BuildSecretVolume(secret)
-			*volumes = appendVolumeIfNotExists(*volumes, volume)
+			*volumes = utils.AppendVolumeIfNotExists(*volumes, volume)
 			container.VolumeMounts =
 				append(container.VolumeMounts, volumeMount)
 			container.Env = append(container.Env,
@@ -121,11 +122,3 @@ func (c *CredentialBuilder) CreateSecretVolumeAndEnv(namespace string, serviceAc
 	return nil
 }
 
-func appendVolumeIfNotExists(slice []v1.Volume, volume v1.Volume) []v1.Volume {
-	for _, ele := range slice {
-		if ele.Name == volume.Name {
-			return slice
-		}
-	}
-	return append(slice, volume)
-}
