+++ pkg/controller/inferenceservice/resources/credentials/service_account_credentials_test.go
@@ -432,123 +432,3 @@ func TestAzureCredentialBuilder(t *testing.T) {
 	g.Expect(c.Delete(context.TODO(), customAzureSecret)).NotTo(gomega.HaveOccurred())
 	g.Expect(c.Delete(context.TODO(), customOnlyServiceAccount)).NotTo(gomega.HaveOccurred())
 }
-
-func TestAppendVolumeIfNotExists(t *testing.T) {
-
-	scenarios := map[string]struct {
-		volumes             []v1.Volume
-		volume      		v1.Volume
-		expectedVolumes 	[]v1.Volume
-	}{
-		"DuplicateVolume": {
-			volumes: []v1.Volume{
-				{
-					Name: gcs.GCSCredentialVolumeName,
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-				{
-					Name: "blue",
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-			},
-			volume: v1.Volume{
-				Name: gcs.GCSCredentialVolumeName,
-				VolumeSource: v1.VolumeSource{
-					Secret: &v1.SecretVolumeSource{
-						SecretName: "user-gcp-sa",
-					},
-				},
-			},
-			expectedVolumes: []v1.Volume{
-				{
-					Name: gcs.GCSCredentialVolumeName,
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-				{
-					Name: "blue",
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-			},
-		},
-		"NotDuplicateVolume": {
-			volumes: []v1.Volume{
-				{
-					Name: gcs.GCSCredentialVolumeName,
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-				{
-					Name: "blue",
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-			},
-			volume: v1.Volume{
-				Name: "green",
-				VolumeSource: v1.VolumeSource{
-					Secret: &v1.SecretVolumeSource{
-						SecretName: "user-gcp-sa",
-					},
-				},
-			},
-			expectedVolumes: []v1.Volume{
-				{
-					Name: gcs.GCSCredentialVolumeName,
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-				{
-					Name: "blue",
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-				{
-					Name: "green",
-					VolumeSource: v1.VolumeSource{
-						Secret: &v1.SecretVolumeSource{
-							SecretName: "user-gcp-sa",
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for name, scenario := range scenarios {
-		volumes := appendVolumeIfNotExists(scenario.volumes, scenario.volume)
-
-		if diff := cmp.Diff(scenario.expectedVolumes, volumes); diff != "" {
-			t.Errorf("Test %q unexpected volume (-want +got): %v", name, diff)
-		}
-	}
-
-
-}
