+++ python/pytorchserver/Dockerfile
@@ -1,0 +1,58 @@
+FROM nvidia/cuda:10.0-cudnn7-devel-ubuntu18.04 AS BUILD
+
+ARG PYTHON_VERSION=3.7
+ARG CONDA_PYTHON_VERSION=3
+ARG CONDA_DIR=/opt/conda
+ARG PYTORCH_VERSION=1.3.1
+
+# Instal basic utilities
+RUN apt-get update && \
+    apt-get install -y --no-install-recommends git wget unzip bzip2 build-essential ca-certificates && \
+    apt-get clean && \
+    rm -rf /var/lib/apt/lists/*
+
+# Install miniconda
+ENV PATH $CONDA_DIR/bin:$PATH
+RUN wget --quiet https://repo.continuum.io/miniconda/Miniconda$CONDA_PYTHON_VERSION-latest-Linux-x86_64.sh -O /tmp/miniconda.sh && \
+    echo 'export PATH=$CONDA_DIR/bin:$PATH' > /etc/profile.d/conda.sh && \
+    /bin/bash /tmp/miniconda.sh -b -p $CONDA_DIR && \
+    rm -rf /tmp/* && \
+    apt-get clean && \
+    rm -rf /var/lib/apt/lists/*
+
+RUN conda install -y python=$PYTHON_VERSION && \
+    conda install -y pytorch==$PYTORCH_VERSION torchvision pillow==6.2.0 cudatoolkit=10.0 -c pytorch && \
+    conda install -y h5py scikit-learn matplotlib seaborn \
+    pandas mkl-service cython && \
+    conda clean -tipsy
+
+# Runtime image
+FROM nvidia/cuda:10.0-base
+
+ARG CONDA_DIR=/opt/conda
+
+# Instal basic utilities
+RUN apt-get update && \
+    apt-get install -y --no-install-recommends git wget unzip bzip2 sudo p7zip && \
+    apt-get clean && \
+    rm -rf /var/lib/apt/lists/*
+
+ENV PATH $CONDA_DIR/bin:$PATH
+ENV CUDA_HOME=/usr/local/cuda
+ENV CUDA_ROOT=$CUDA_HOME
+ENV PATH=$PATH:$CUDA_ROOT/bin:$HOME/bin
+ENV LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$CUDA_ROOT/lib64
+
+RUN mkdir -p /opt/conda/
+
+WORKDIR /workspace
+RUN chmod -R a+w /workspace
+
+COPY --chown=1000 --from=build /opt/conda/. $CONDA_DIR
+COPY pytorchserver pytorchserver
+COPY kfserving kfserving
+
+RUN pip install --upgrade pip && pip install -e ./kfserving
+RUN pip install -e ./pytorchserver
+ENTRYPOINT ["python", "-m", "pytorchserver"]
+
